Image spacing to make MkDocs/GitHub happy
Whooo, different flavors of Markdown....

Handy reference for just how ridiculous the universe is: http://johnmacfarlane.net/babelmark2/?text=1.+list%0A2.+second%0A++!%5Bimage%5D(image.png)%0A%0A3.+third%0A%0A

I did some testing, and here's how I *think* are some simple rules that keeps MkDocs and GitHub happy:

- All images should have a blank line before and after it.
- If it is not part of a list, do *not* indent it.
- If it is part of a list, indent *exactly* four spaces.

That's it -- if we lose a blank line here or there, or the spacing is off, it might break one or the other, or it might not matter at all, but I wanted to avoid complicating the rules as much as possible. Even then, due to the ambiguous nature Markdown syntax, we might *still* run into a weird issue, so... but I hope this is clear enough for now.